refactor: 패키지 변경 및 생성자에서 검증 추가
