+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -12,11 +12,14 @@ import jakarta.persistence.ManyToOne;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Objects;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 import roomescape.member.domain.Member;
 import roomescape.theme.domain.Theme;
 import roomescape.time.domain.ReservationTime;
 
 @Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Reservation {
 
     @EmbeddedId
@@ -41,8 +44,8 @@ public class Reservation {
     @JoinColumn(name = "theme_id", nullable = false)
     private Theme theme;
 
-    protected Reservation() {
-    }
+//    protected Reservation() {
+//    }
 
     public Reservation(
             final Long id,
