+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -24,15 +24,15 @@ import roomescape.payment.dto.request.PaymentRequest;
 import roomescape.payment.service.PaymentService;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationId;
-import roomescape.reservation.domain.ReservationWithPayment;
 import roomescape.reservation.dto.request.FilteringReservationRequest;
 import roomescape.reservation.dto.request.ReservationCreateRequest;
 import roomescape.reservation.dto.response.BookedReservationTimeResponse;
 import roomescape.reservation.dto.response.MyReservationsResponse;
 import roomescape.reservation.dto.response.ReservationResponse;
+import roomescape.reservation.dto.response.ReservationWithPayment;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservation.waiting.domain.Waiting;
-import roomescape.reservation.waiting.domain.WaitingWithRank;
+import roomescape.reservation.waiting.dto.response.WaitingWithRank;
 import roomescape.reservation.waiting.repository.WaitingRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.ThemeId;
