+++ src/main/java/roomescape/reservation/waiting/dto/response/WaitingWithRank.java
@@ -1,0 +1,10 @@
+package roomescape.reservation.waiting.dto.response;
+
+import jakarta.validation.constraints.NotNull;
+import roomescape.reservation.waiting.domain.Waiting;
+
+public record WaitingWithRank(
+        @NotNull Waiting waiting,
+        @NotNull long rank
+) {
+}
